requirements.txt
discord.py
aiohttp
pyyaml
pycryptodome
